+++ footer.php
@@ -6,7 +6,9 @@
 
 	</div><!-- .site-content -->
 
-  <?php the_telabotanica_module('footer', []); ?>
+	<?php
+	if ( !isset($GLOBALS['is_error']) ) the_telabotanica_module('footer');
+	?>
 
 <?php wp_footer(); ?>
 </body>