+++ CHANGELOG.md
@@ -33,6 +33,7 @@ plus de détails sur http://semver.org/lang/fr/
 	* support de multiples pages parentes
 * module `cover-home` : ajout du lien vers le tableau de bord
 * module `cover-member` ajouté
+* module `error` ajouté
 * module `feed` ajouté
 * module `feed-date` ajouté
 * module `feed-item` ajouté
@@ -58,6 +59,7 @@ plus de détails sur http://semver.org/lang/fr/
 * template `page-comment-participer` ajouté
 * ajout d'une taxonomie Catégorie de moyens de participer
 * groupe de champs ACF "Composant liste d'articles" ajouté
+* pages d'erreur 404, 500, maintenance
 
 ## 0.0.7 (2017-03-04)